[compress] move knet_handle_get_compress_list to knet_get_compress_list
now matches the same API as get_transport_list
