+++ libknet/libknet.h
@@ -802,14 +802,14 @@ struct knet_crypto_info {
  *
  * @brief Get a list of supported crypto libraries
  *
- * crypto_list  - array of strict knet_crypto_info *
+ * crypto_list  - array of struct knet_crypto_info *
  *                If NULL then only the number of names is returned in crypto_list_entries
  *                to allow the caller to allocate sufficient space.
  *		  libknet does not allow more than 256 crypto methods at the moment.
  *		  it is safe to allocate 256 structs to avoid calling
  *		  knet_get_crypto_list twice.
  *
- * num_names - returns the number of strings in crypto_names
+ * crypto_list_entries - returns the number of strings in crypto_names
  *
  * @return
  * knet_get_crypto_list returns
@@ -820,26 +820,34 @@ struct knet_crypto_info {
 int knet_get_crypto_list(struct knet_crypto_info *crypto_list,
 			 size_t *crypto_list_entries);
 
+
+
+struct knet_compress_info {
+	const char *name; /* bzip2, lz4, etc.. */
+};
+
 /**
- * knet_handle_get_compress_list
+ * knet_get_compress_list
  *
  * @brief Get a list of support compression types
  *
- * knet_h   - pointer to knet_handle_t
+ * compress_list - array of struct knet_compress_info *
+ *		   If NULL then only the number of names is returned in num_names to
+ *		   to allow the caller to allocate sufficient space.
+ *		   libknet does not allow more than 256 compress methods at the moment.
+ *		   it is safe to allocate 256 structs to avoid calling
+ *		   knet_get_compress_list twice.
  *
- * compress_names - array of char *
- *                If NULL then only the number of names is returned in num_names to
- *		  to allow the caller to allocate sufficient space.
- *
- * num_names - returns the number of strings in compres_names
+ * compress_list_entries - returns the number of strings in compres_names
  *
  * @return
- * knet_host_add returns
+ * knet_get_compress_list returns
  * 0 on success
  * -1 on error and errno is set.
  */
 
-int knet_handle_get_compress_list(knet_handle_t knet_h, const char **compress_names, size_t *num_names);
+int knet_get_compress_list(struct knet_compress_info *compress_list,
+			   size_t *compress_list_entries);
 
 /*
  * host structs/API calls
