+++ libknet/tests/api-check.mk
@@ -20,7 +20,7 @@ api_checks		= \
 			  api_knet_handle_get_datafd_test \
 			  api_knet_handle_get_stats_test \
 			  api_knet_get_crypto_list_test \
-			  api_knet_handle_get_compress_list_test \
+			  api_knet_get_compress_list_test \
 			  api_knet_handle_clear_stats_test \
 			  api_knet_get_transport_list_test \
 			  api_knet_get_transport_name_by_id_test \
@@ -110,7 +110,7 @@ api_knet_handle_get_stats_test_SOURCES = api_knet_handle_get_stats.c \
 api_knet_get_crypto_list_test_SOURCES = api_knet_get_crypto_list.c \
 					test-common.c
 
-api_knet_handle_get_compress_list_test_SOURCES = api_knet_handle_get_compress_list.c \
+api_knet_get_compress_list_test_SOURCES = api_knet_get_compress_list.c \
 					  test-common.c
 
 api_knet_handle_clear_stats_test_SOURCES = api_knet_handle_clear_stats.c \
